+++ app/src/main/java/org/simple/clinic/registration/facility/RegistrationFacilitySelectionUpdate.kt
@@ -4,7 +4,9 @@ import com.spotify.mobius.Next
 import com.spotify.mobius.Update
 import org.simple.clinic.mobius.dispatch
 
-class RegistrationFacilitySelectionUpdate : Update<RegistrationFacilitySelectionModel, RegistrationFacilitySelectionEvent, RegistrationFacilitySelectionEffect> {
+class RegistrationFacilitySelectionUpdate(
+    private val showIntroVideoScreen: Boolean
+) : Update<RegistrationFacilitySelectionModel, RegistrationFacilitySelectionEvent, RegistrationFacilitySelectionEffect> {
 
   override fun update(
       model: RegistrationFacilitySelectionModel,
@@ -12,11 +14,21 @@ class RegistrationFacilitySelectionUpdate : Update<RegistrationFacilitySelection
   ): Next<RegistrationFacilitySelectionModel, RegistrationFacilitySelectionEffect> {
     return when (event) {
       is RegistrationFacilityClicked -> dispatch(OpenConfirmFacilitySheet(event.facility))
-      is RegistrationFacilityConfirmed -> {
-        val updatedEntry = model.ongoingEntry.withFacilityUuid(event.facilityUuid)
-
-        dispatch(MoveToIntroVideoScreen(updatedEntry))
-      }
+      is RegistrationFacilityConfirmed -> registrationFacilityConfirmed(model, event)
     }
+  }
+  
+  private fun registrationFacilityConfirmed(
+      model: RegistrationFacilitySelectionModel,
+      event: RegistrationFacilityConfirmed
+  ): Next<RegistrationFacilitySelectionModel, RegistrationFacilitySelectionEffect> {
+    val updatedEntry = model.ongoingEntry.withFacilityUuid(event.facilityUuid)
+    val effect = if (showIntroVideoScreen) {
+      MoveToIntroVideoScreen(updatedEntry)
+    } else {
+      MoveToRegistrationLoadingScreen(updatedEntry)
+    }
+
+    return dispatch(effect)
   }
 }